% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyssd_run.R
\name{shinyssd_run}
\alias{shinyssd_run}
\title{shinyssd: species sensitivity distributions for ecotoxicological risk assessment}
\usage{
shinyssd_run()
}
\description{
The shinyssd web application is a versatile and easy to use tool that serves to simultaneously model the species sensitivity distributions (SSD) curve of a user-defined toxicity dataset based on four different statistical distribution models (log-normal, log-logistic, Weibull, Pareto). shinyssd directly calculates three estimators Hazard Concentration 1 percent (HC1), 5 percent (HC5) and 10 percent (HC10) associated to the four distribution models together with its confidence intervals, allowing the user to select the statistical distribution and associated HC values that best adjust the dataset.
}
\details{
Start shinyssd
}
\examples{
\dontrun{
library(shinyssd)
shinyssd::shinyssd_run()
}
}
\keyword{shiny}
\keyword{shinyssd,}
\keyword{ssd,}
